Refining animation and adding links
